test/requirements.txt
harness-featureflags
requests
fastapi
uvicorn[standard]